==> notebooks/time/mouse_atlas/2_Supplementary_figures/5_Metacells/0_Select_single_timepoint_anndatas.ipynb <==
import numpy as np
import anndata
import pandas as pd
import scanpy as sc
import scipy
import seaborn as sns
import matplotlib.pyplot as plt
import os
import time

ts0 = 'E9.5'
ts1 = 'E10.5'

Path="/home/icb/manuel.gander/mouse_atlas/data"
adata0=sc.read(f"{Path}/Comb_anndatas/adata_{ts0}_{ts1}.h5ad")
del adata0.raw
adata0.obs['day']=adata0.obs['day'].astype('category')
a0 = adata0[adata0.obs['day']==9.5].copy()
a0.write('E95_full.h5ad')

ts0 = 'E10.5'
ts1 = 'E11.5'
Path="/home/icb/manuel.gander/mouse_atlas/data"
adata0=sc.read(f"{Path}/Comb_anndatas/adata_{ts0}_{ts1}.h5ad")
del adata0.raw
adata0.obs['day']=adata0.obs['day'].astype('category')
a0 = adata0[adata0.obs['day']==10.5].copy()
a1 = adata0[adata0.obs['day']==11.5].copy()
a0.write('E105_full.h5ad')
a1.write('E115_full.h5ad')
