==> notebooks/time/mouse_atlas/2_Supplementary_figures/5_Metacells/4_Cell_type_transition_accuracies.ipynb <==
import sys,os
sys.path.append('/home/mgander/mouse_atlas')
import c2
import pandas as pd

CT0 = pd.read_pickle('/home/mgander/mouse_atlas/data/moscot_maps/Anchors/CTs/E10.5_E11.5_cell_type_transitions.pkl')

c2.evaluate_using_curated_transitions(CT0, cutoff=0.00)

c2.evaluate_using_germ_layers(CT0, cutoff=0.00)

CT1 = pd.read_pickle('CT_metacells2.pkl')
CT1 = CT1/CT1.sum()

c2.evaluate_using_curated_transitions(CT1, cutoff=0.00)

c2.evaluate_using_germ_layers(CT1, cutoff=0.00)

vals = [0.779983, 0.964083, 0.756499, 0.956559]
method = ['single_cell', 'single_cell', 'metacell', 'metacell']
eval_type = ['curated', 'germ', 'curated', 'germ']
pd.DataFrame({'vals':vals, 'method':method, 'eval_type':eval_type})

pd.DataFrame({'vals':vals, 'method':method, 'eval_type':eval_type})
